refactor(web): change directory structure
