+++ apps/web/src/pages/home/ui/index.ts
@@ -1,4 +1,1 @@
 export { default as HomePage } from './HomePage';
-export { default as CreateProjectModal } from './CreateProjectModal';
-export { default as LogoutModal } from './LogoutModal';
-export { default as SettingModal } from './SettingModal';
