+++ apps/web/src/widgets/home/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeContainer } from './HomeContainer';
