+++ apps/web/src/widgets/modal/ui/index.ts
@@ -1,0 +1,4 @@
+export { default as CreateProjectModal } from './CreateProjectModal';
+export { default as LogoutModal } from './LogoutModal';
+export { default as SettingModal } from './SettingModal';
+export { default as AlarmModal } from './AlarmModal';
